\name{geigen-package}
\alias{geigen-package}
\alias{geigen.Intro}

\title{\packageTitle{geigen}}

\description{\packageDescription{geigen}}
\details{
The \pkg{geigen} package provides functions to compute the generalized eigenvalues of a pair of matrices and
optionally provides the generalized eigenvectors. Both numeric and complex matrices are allowed.
The package also provides a function for computing the generalized Schur decomposition of a pair of matrices,
either numeric or complex.
Finally the package provides a function for computing the generalized singular value
decomposition for a pair of matrices, either numeric or complex.
The package uses routines provided by the Lapack linear algebra package.
}

\keyword{package}
