\name{geigen-package}
\alias{geigen-package}
\alias{geigen.Intro}
\docType{package}
\title{Calculating Generalized Eigenvalues}
\description{
The \pkg{geigen} package provides functions to compute the generalized eigenvalues for a pair of matrices and
optionally provides the generalized eigenvectors. Both numeric and complex matrices are allowed.
The package also provides a function for computing the generalized Schur decomposition of a pair of matrices,
either numeric or complex.
% The package uses Lapack subroutines and calls these with the .Fortran interface.
}
% \details{
% \tabular{ll}{
% Package: \tab geigen\cr
% Type: \tab Package\cr
% Version: \tab 0.9\cr
% Date: \tab 2013-05-05\cr
% License: \tab GPL-2\cr
% }
% ~~ An overview of how to use the package, including the most important ~~
% ~~ functions ~~
% }

\author{Berend Hasselman \email{bhh@xs4all.nl}}

% \references{
% % ~~ Literature or other references for background information ~~
% }
% ~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
% ~~ the R documentation directory ~~
\keyword{package}
% \seealso{
% % ~~ Optional links to other man pages, e.g. ~~
% % ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
% }
% \examples{
% }
